+++ plugin/www/applovinmax.js
@@ -286,6 +286,9 @@ window.addEventListener('OnRewardedAdDisplayedEvent', function (adInfo) {
 window.addEventListener('OnRewardedAdFailedToDisplayEvent', function (adInfo) {
     AppLovinMAX.isAdReadyValues[adInfo.adUnitId] = false;
 });
+window.addEventListener('OnRewardedAdReceivedRewardEvent', function (adInfo) {
+    AppLovinMAX.isAdReadyValues[adInfo.adUnitId] = false;
+});
 
 if (typeof module !== undefined && module.exports) {
     module.exports = AppLovinMAX;