+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Versions
 
-## 1.2.0
+## x.x.x
     * Added "OnRewardedAdReceivedRewardEvent" listener.
 ## 1.1.5
     * Depends on Android SDK [v12.3.1](https://github.com/AppLovin/AppLovin-MAX-SDK-Android/releases/tag/release_12_3_1) and iOS SDK [v12.3.1](https://github.com/AppLovin/AppLovin-MAX-SDK-iOS/releases/tag/release_12_3_1).