+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Replace deprecated `Targeting Data` APIs with `Segment Targeting`.
     * Remove deprecated `ConsentDialogState` APIs.
     * Remove deprecated `isAgeRestrictedUser` APIs.
+    * Add a listener for `OnRewardedAdReceivedRewardEvent`.
 ## 1.1.5
     * Depends on Android SDK [v12.3.1](https://github.com/AppLovin/AppLovin-MAX-SDK-Android/releases/tag/release_12_3_1) and iOS SDK [v12.3.1](https://github.com/AppLovin/AppLovin-MAX-SDK-iOS/releases/tag/release_12_3_1).
 ## 1.1.4